feat: Pass --use-feature='dynamic-feed' for new CLI
